requirements.txt
imblearn==0.0
joblib==1.4.2
lazypredict_nightly==0.3.2
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.5.2
seaborn
timezonefinder==6.5.3
tqdm>=4.66.3
ydata_profiling
